Display attachments, Refactor
